Update filecp
This script uses the find command to search for files with the specified media file types in the source directory, and then copies them to the destination directory using the cp command.

You can modify the types array to include any other file types that you want to copy. Just make sure to include the file extension in the array.